refactor: move plug response delivery into runtime
